dcache-view: bind page refresh to routing event
Motivation:

Many of the admin view pages and dialogs use setInterval to
refresh the data automatically.  In terms of deactivation,
the dialogs currently behave correctly because the timer
is explicitly stopped when the dialog is closed.  For the
page views, however, the refresh continues once the page
is loaded, regardless of whether the page is the current
routing target or not.  This is suboptimal, as there is
no need to refresh data for invisible pages.

Modification:

A window event, dv-routing-changed, is now propagated from the
appropriate place on the routing page.  The mixin
implements the necessary logic for turning the refresh
on and off.  Each non-dialog component which uses
refresh now sets its component path internally so
that a validation check can be made.  Refresh
occurs only if the valid flag is true.

Result:

When switching between views, refresh is cancelled
for all but the page which is visible, if it uses
refresh.

Target: master
Request: 1.4
Issue: https://github.com/dCache/dcache-view/issues/91
Acked-by: Olufemi