dcache-view (authentication): fix admin role assertion for showing admin page
Motivation:

Roles can be asserted in dcache-view in one of these three ways, by:
 a. using the toggle botton on the user-profile page (if the user had
  already been authenticated);
 b. typing `<username>#<comma-seperated-roles-list>` in the username
  textbox of the login page; and
 c. setting the checkbox for "Assert all roles" to true.

Since the admin page will only be visible to user with admin role, hence
when user assert the admin role by using the option two, this is not
the case.

Modification:

Listen to `dv-authentication-successful` event in the dv.js and response
accordingly when admin role is asserted.

Result:

Admin page will be accessible when a admin role is asserted and
irrespective how it is asserted.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Paul Millar
Acked-by: Albert Rossi

Reviewed at https://rb.dcache.org/r/11091/