dcache-view (authentication): fix openid connect redirect handling
Motivation:

When an openid connect provider redirect back to dcache-view, the
access_token needs to be extracted and verified. At the moment,
before the extraction of the access_token, dcache-view checked if
the redirect url contain some other parameters which might be
problematic based on the provider response.

Also, when error occuur during the verification process of the
access_token, the was not properly handled because the name of
the event had been changed without updating this part.

Modification:

1. check if the redirect url contain only access_token since this
the only important parameter for dcache-view.
2. filer the redirect url properly to extract the only needed
access_token.
3. change the name of the `error` event listener to
`dv-authentication-error`, which is the latest name.

Result:

- Proberly handle error that might occur during verification of
access_token in dcache.

- Extract only the needed parameter from the redirect url.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11125/