dcache-view (authentication): minor fixes in the oic-login-form page
Motivation:

Motivation:

1. If there are more than one oic providers, the json array that
suppose to form an array list based on the information from the
config file will thrown an error because the object are not
properly created.

2. When a css overflow value is set to scroll, some broswer will
not show any scroll bar if there is nothing to scroll, while
others by default will show the scrollbar. Due to this inconsident
in behaviour, the rendering of this page unharmonious across
different platform.

3. The oic provider buttons are created base on the number of
the providers known to dcache-view. At times this buttons are
unresponsive when clicked because the data are not set and data-\*
attribute behave inconsistently with dom-repeat template.

4. In the request url, dcache-view always request that the oic
provider should always request for user consent. This is
undesirable especially when the user had already consented.

Modification:

1. Set the css overflow to auto.
2. remove all inline styling
3. move the setting of oic providers to the constructor phase
and use es6 for forming the object for each provider.
4. use `itemForElement` method that come with dom-repeat
template to get the values of the providers instead of using
data-attribute.

Result:

1. Better styling and uniform interface across different platform
2. provider button now works as espected.
3. no more annoying user consent page.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11126/