dcache-view: improve handling of mover kill
Motivation:

When multiple movers are killed, it is best to wait until
all have been processed before attempting to refresh.

Also, somewhat more specific error messages as to the
cause of failure should be displayed.

Modification:

Store the urls and delete them upon response; refresh
when the list is empty.

Set the error handling to use reject with request,
so that the returned error from the server is made
available.  Extract all errors from it for display.

Result:

Smoother handling of refresh and somewhat more
informative error messages.

Target: master
Request: 1.4
Acked-by: Olufemi