dcache-view (user-profile): fix gravatar request
Motivation:

Since the upgrade to polymer 2.x elements which depends on
the node dom must be explict append. This is not case with
gravatar-request element and since it is not attached to
the dom tree. iron-jsonp-request element will not fire
even when the library url is set.

Modification:

1. Redesign how the gravatar-request element work, by:
    - requiring no set-up at constructor phase;
    - making the element hidden by default, since is main
    function is just to make a request. Hence, it can be
    added to the dom tree and behave as such but will
    not be visible to the user;
    - ensuring a request for gravatar was be made when
    the email property is set or when the request method
    is used.
    - requesting for another gravatar if the first request
    failed and more than one emails were provided.
    - removing all the unneccessary properties

2. Adjust the user-image to use the redesign gravatar-request
element. Hence, the gravatar-request element is part of its
children and annotated listeners for both successful and
unsuccessful response were attached.
Also worth noting is the fact that if for whatever reason
no response is received from the gravatar-request element,
the fallback; identicon will be use.

3. remove the size attribute in the user-image tag inside
the user-profile, since this is the default value of the
user-image element. Add a fallback-value attribute to the
user-image tag. And the attribute is set to the username.

Result:

1. The identicon will be shown if gravatar request is
unsuccessful or no response from the element over a set
period of time. This ensure that the loading icon will
not be display infinitely.
2. The gravatar request is definite made when user email
is provided.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: ALbert Rossi
Fixes: https://github.com/dCache/dcache-view/issues/99

Reviewed at https://rb.dcache.org/r/11135/