dcache-view (namespace): remove restriction on delete operation
Motivation:

At the moment, delete opereation will not be process if
dcache-view thinks the user is anonymous, that is, the
request will not be send to the server.

This is not a desirable result since for two reasons:
- client might have a valid certificate that dcache had
already authenticated or
- the site might permit anonymous operation on delete.

Modication:

- Remove restriction on delete.
- Use arrow function where necessary

Result:

The delete operation is always sent to the server.

Target: master
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11159/