dcache-view (namespace): fix style issue in list-row & list-row-header
Motivation:

dCache View uses list-row element to represent and show a
file in a list. It contain 5 columns: file icon, file name,
creation time, file location and file size.

But there are two problems with list-row element:

1. When the screen size is small, all the 5 columns are still
shown even when the client screen width is not sufficient.
2. If the text length of each name of all the files in
a directory were greatly varies from each other; this
will result in the distortion of the uniformity and the
consistency that should be maintained for each columns in
the list.

Modification:

For both list-row and list-row-header:

1. use media queries to create flexible layout for list-row
element. If the screen is less than 960px, only two columns
(which are the file icon and file name colums) will be shown.

2. set the columns to a particular fixed width and make the
file name column flexible but with a set max-width. The
max-width of the file name column is based on the size of the
screen. Hence, if the text length of file name is greater than
the max-width assigned, the css text-overflow and text-overflow
properties of the column will ensure that text is trunctated
with an appended ellipsis.

Target: master
Request: 1.4
Request: 1.3
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11174/

(cherry picked from commit 06a50f11c52ea7403a64b0929471c1ae1ac41e55)