dcache-view (namespace): fix header and preview display
Motivation:

When the length of the text in the title bar of the
file-metadata-dashboard element is more than the available
space, the header become unfitted and out if place. One of
the main reasons is that the css property word-wrap is set
to break-word, hence making the text container space height
much more than the header height.

Also, the preview display have a smillar problem, whenever
the file name is too long.

Modification:

1. use css property white-space and it is set to nowrap
    and add text-overflow and it is set to ellipsis. Setting
    both properties will ensure that if the header text
    length is more the available space, an ellipsis will
    be shown instead of displaying all text.
2. place the content of the previow inside the p tag. And
    the restriction on the height is removed. This enable
    that all the text are shown in an orderly fashion no
    matter how long they are.

Result:

Better display to the end user.

Target: master
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11173/