dcache-view (namespace & file-viewing): discover webdav doors
Motivation:

At the present, it required webdav door's url be set in other
for dCache View to communicate with the webdav doors. If no
value is set the default is assumed. Since there is a newly
added API in dCache, which one can use to get or obtain list
of available doors, it become neccessary to auto discover the
webdav doors.

Modification:

1. Create a worker to obtain list of available doors and narrow
it down to only webdav doors. Rank the webdav doors based on the
protocol use by dCache View and the load time property of the
door. Also, the webdav doors are split into two based on write
and read.

2. Use the worker and store the result in the window.CONFIG.webdav

3. Create a method in the util mixin class that will take in a
path and the operation type to return an array containing list
of url that can be used to perform the necessary operation.

4. Adjust both file-viewer and upload section to use the mixin
method for get the webdav url with the path.

Result:

Less properties to set by the admin.

Target: master
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11270/