dcache-view (namespace): use webworker to fetch qos
Motivation:

JavaScript code is single-threaded, hence all non DOM
computation and manipulation should be done inside a
worker. Depending on the file size, changing the QoS
of a file might be time consuming operation. Therefore,
this operation is a very good candidate for background
thread task.

Also, when a change of QoS is requested, an arrow
showing the target QoS suppose to be shown but somehow
the unicode character added to the target QoS make it
display `undefine`.

Modification:

- Perform the recursive request for getting the file
current QoS status in the worker.
- Use real arrow icon that will be shown when the file
is in transition. This arrow will be blinking and after
the transition is complete, it will be remove from the
dom.

Result:

Show correct transition status to the user and this free
the main thread of blocking operation.

Target: master
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan

Reviewed at https://rb.dcache.org/r/11275/