dcache-view (authentication): update getAuthValue method
Motivation:

We now support certificate for authourisation and authentication,
soon we will be accepting macaroons for the same thing as well.
Hence, adjustment is needed on how we set and get authourisation
value.

Modification:

Update the `apps.getAuthValue` method to take in an object argument.
If an argument is supplied, this will be used even when client
certificate is present, else client certificate is checked or the
legacy code is fallback on.

Result:

It provide a single point to set and get authourisation value.

Target: master
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11276/