dcache-view (namespace): fix notification in dnd move op
Motivation:

When a file is dragged and dropped, a move operation
is initiated. A snackbar is shown to provide the user
a feedback on initiated operation.

Unfortunately, this is broken. For instant, when an
error occured, a successful move message was shown.

Modification:

Send message after each move.

Result:

User now see a correct message when a move operation
is initiated through drag and drop.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan

Reviewed at https://rb.dcache.org/r/11278/