dcache-view (authentication): adjust elements to use getAuthValue()
Motivation:

Some elements are written in such a way that they always uses the base
authentication, that is, the user signed in authenticated credentials.
This is not desirable in some suitation like when the known user have
a macaroon to perform some operation that s/he wouldn't have had the
permission to.

Modification:

- Add in the common mixin a property called `authenticationParameters`
and a method that will set and get authentication value base on the
`authenticationParameters`. If this parameter is not defined, a fall-
back option will be returned.

- Update some elements to use the common mixin and if an element already
uses the mixin, the is adjuested to use the new property and method.

- Update the download-task script (which is a worker) to use the returned
value from the getAuthValue method. Since we know that an empty string
will be returned by the method, if an authenticated client certificate
is present, hence there is no need for `e.data.usesCert`.

Result:

Allow elements to use different auth values for different request.

Target: master
Require-notes: no
Require-book: no
Acked-by: Albert Rossi

Reviewed at https://rb.dcache.org/r/11305/