dcache-view: fix computation of layout bar percentages
Motivation:

Pool and pool group space data is returned as raw
numbers in the json delivered by the RESTful API,
but is displayed in two ways on the pool group view
and pools dialogs:  as a text amount (e.g., total = 4.3 TB)
and as a layout bar with colored portions corresponding
to free, precious and removable files on the pool(s).

At some point in the development, the function which
parses the raw data into an internal data object
was changed to represent the totals using byte-multiple
suffixes.  However, the change neglected to preserve
the raw numbers for use in calculating the percentages
on the layout bar markup.

Modification:

Separate out the data fields used by the text vs
the layout bar.

We have also added a small function to common.html
to check for number, and called this in the approprate
places so that NaN is simply treated as 0.

Result:

Text displays are intelligible and the layout bar
composition is correct.

Target: master
Request: 1.4
Acked-by: Olufemi