dcache-view: alter the way multiple errors are handled
Motivation:

When errors occur on HTTP requests, they are usually handled
in the admin interfaces by passing them to a 'toast' component
for display.   This usually is fine when an error associated
with a single request is involved.

However, when there are multiple asynchronous requests involved,
the direct reporting of each one using the 'toast' is a bad
idea, for two reasons:

(A)  if there are multiple such errors, the display of the toast
     blocks the rendering thread and as such the other errors
     get lost;

(B)  even a single such asynchronous error can sometimes affect
     the processing of the remaining requests, stopping them
     from completing.

Modification:

The display of errors messages in dCache-View in general needs
to be rethought; but in the meantime, the places where the toast
could be called within the context of multiple asynchronous
requests have been altered such that the any such errors are
appended to a single message which is displayed only when
all such requests complete.

Result:

Normal operations should complete despite of asynchronous
error handler calls, and the message delivered to the user
should have information that reflects the multiple error
instances.

Target: master
Request: 1.4
Acked-by: Paul