dcache-view (gulp-tasks): ensure copy-pdfjs task successful execution
Motivation:

We need the minified version of pdfjs library and the build
needs to be in a proper order in dcache-view task list so
that the copy-pdfjs task will be executed successfully.

Modification:

- deep copy the build directory of pdfjs folder
- move the task towards the end of all tasks.

Result:

The minified version of pdfjs is included in the successful
build.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11362/