dcache-view (namespace): create worker for move operation
Motivation:

The move operation request is one of those tasks that does
not directly involve in DOM manipulation. Hence, a perfect
candidate for a worker. The will result in making dCache-
view more responsive and the rename-input element become
even more modular.

Modification:

- Create a worker for namespace move operation. The operation
uses fetch API to send the request to dCache.This worker can
be use to either move a file or rename it.

- Update the rename-input element to use the worker and set
the authentication value at the constructor level if it is
provided.

- Adjust view-file element to use the updated rename-input
element.

Result:

Better user experience. Also, a noticed bug (which is: when
a user request is unsuccessful, the file name was updated to
the new name) is now fixed with this patch.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11383/