dcache-view (user-profile): make gravatar optional
Motivation:

If a user account have an email associated with it, a request
is sent to check if the email is registered with the Gravatar,
and if it is, the image from gravatar is used by dcache-view
as the user profile.

Since this is the current default behaviour of dcache-view,
some sites are not happy with this and the preferred behaviour
will to make gravatar optional for users.

Modification:

1. properly remove all node inside the user profile section.
2. add a checkbox in the login form. This enable users to
    choose whether to use a gravatar or not. By default,
    the checkbox is unchecked. This means that users will
    have to specfically check the checkbox to indicate that
    dcache-view can make a request to the Gravatar
3. adjust the user-image element
4. update user-profile and user-profile-dropdown element to
    use the adjusted user-image element.
5. add a section in the user-profile for user to see and sel-
    ect the preferred user profile image. There are only two
    options (identicon and gravatar) available at this time.

Result:

User can now opt-in or out of gravatar.

Target: master
Request: 1.5
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan
Fixes: https://github.com/dCache/dcache-view/issues/135

Reviewed at https://rb.dcache.org/r/11450/