dcache-view (file-sharing): add support for macaroon request
Motivation:

One of the possible usage of the new added bearer token
called Macaroon in dcache is for sharing files. To allow
dcache-view user to be able to share files, this patch
add support for macaroon request.

Modification:

1. add a share file action in the context menu
2. create a file-sharing-form element which user will use
to restrict macaroon permission.
3. create a script to request for macaroon bear token
and to generate a sharing link for dcache-view.
4. use a 3rd party library for generating QR code for
the generated link.

Result:

1. Genereate a shareable link that can be copied by the user.
2. The generated link is use to create a QR code for the
shareable link that can be downloaded to the user disk.

Note: This is for requuest a shareable link or macaroon.
The viewing of shared files will be in a different patch.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Acked-by: Albert Rossi

Reviewed at https://rb.dcache.org/r/10953/