dcache-view (upload): fix dcache issue 3011
Motivation:

When a user upload a file and the re-uploaded it (for whatever
reason), the browser always prevent this kind of re-upload.
This is mainly because the value of the input element have
not changed and the on-change event is not triggered.

Modification:

Set the input value to null when the tag is clicked.

Result:

Fixed the reported issue.

Target: master
Request: 1.5
Request: 1.4
Request: 1.3
Require-notes: no
Require-book: no
Acked-by: Paul Millar
Fixes: https://github.com/dCache/dcache/issues/3011

Reviewed at https://rb.dcache.org/r/11434/