dcache-view (upload): fix issue 113
Motivation:

Sometimes if user doesn't have the right permission or the
webdav is not trusted by the browser. If this user perform
a write operation with dcache-view upload handler, the ajax
request will return an empty `statusText` with zero `Status`
code number but with `readyState` set to 4. This implied that
the request was sent and error occured but with no error me-
ssage. For the end-user, this is not very helpful.

Modification:

- adjust upload-toast-file, by first using the one-way
    binding. And second, when the error message is empty,
    a button will be shown. If clicked, this will show a
    dialog box that contain information concerning the
    possible cause of the error.
- create an element that will serve as a dialog content
    for the upload failure.

Result:

Provide some possible cause of the upload failure.

Target: master
Request: 1.5
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Albert Rossi
Fixes: https://github.com/dCache/dcache-view/issues/113

Reviewed at https://rb.dcache.org/r/11440/