dcache-view (file-sharing & user-profile): redesign user-image
Motivation:

The user-image element is used in multiple places within dcache-view.
Initially, this element was designed to be used in the user profile
page. Recently, the internal working of user-image was change to give
end users the preference to choose whether to use gravatar or not.
But the file-sharing usage of user-image was not accommodated in the
previous redesign.

Also, all user-images created pointed to a single endpoint and if
there is a switch between the user prefernce, a new image source is
created. This restricts multiple usage of user-elements for different
purpose and unnecessary request are sent each time user changes it
preferences.

Modification:

1. update user-image inside the shared-file-list element to the new
    new design also to have a fixed size.
2. adjust the dispatch custom element inside the user-profile for
    image request change to contain more details. Also, update user
    -image node to the new design.
3. update user-image node inside the user-profile-dropdown element.
3. remove setting of image endpoint inside the gravatar-request
    element.
4. redesign user-image to be able to:
    - store and update different endpoints based on some of the
    properties. This will be referred to later when the user
    preference matched the stored values.
    - make each created user-image have an identifier. This
    identifier can be use to control or adjust a bunch user-images
    irrespective of where they are.

Result:

- Make user-image useable in file-sharing
- manage and set different endpoints for all the user-images
    created.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11470/