dcache-view (filemetadata-dashboard): show correct information
Motivation:

When a directory listing is requested and if an error ocurred
(this might be error due to client side or server side error),
this is usually display on the directory listing page. If a
user should open the file-metadata dashboard (either through
context menu or by clicking the file metadata button at the top
right corner of dcache-view page), a misleading information is
show, which doesn't reflect the error shown in the directory
listing page.

Modification:

Add a message to the data of the custom event object dispatched
(the event is called `dv-namespace-receive-children`) in view-file
element. This message is set based on the response of the server
regarding the request of currently viewed directory. This enable
file-metadata-dashboard element (which listen to this event) to
display appropriate information corresponding to currently viewed
directory.

Result:

Correct metadata information is shown.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan

Reviewed at https://rb.dcache.org/r/11478/