dcache-view (npm-package): add bower to list of dependencies
Motivation:

For anybody to build dcache-view from scratch, this requires that
the user independently install maven and bower. Although the number
of dependencies needed by dcahe-view is more than those two but
these dependencies have been added to the one that will be instal-
led by npm. By using a maven plug-in, npm and node are installed
automatically and locally. Hence, making it easy to install any
third party dependency.

To install bower, this requires node and npm. Since bower is one
of the required dependencies for dcache-view and the maven plugin
already have node and npm running locally for the project, it is
not neccessary to install this bower independently.

Modification:

Add bower to the list of dependencies that will be required in the
package.json file.

Result:

Make dcache-view much more easy to build by reducing the amount of
independent dependencies required to only maven.

Target: master
Request: 1.5
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/11485/