dcache-view: fix label for time in mover listing
Motivation:

The value is in milliseconds, not seconds.

The API has been changed, so the field name for sorting
and the object property needs to be changed here.

Modification:

Change timeInSeconds to timeInMilliseconds.

Result:

Sorting and display of value will work properly after
update of dcache master and branches.

Target: master
Request: 1.5
Request: 1.4
Acked-by: Olufemi
Acked-by: Dmitry