dcache-view (download): show useful error message
Motivation:

If an error occur when a user request to download a file,
the correct error message is not shown to the user.

Modification:

Check if the response for the request is succeful and if
not propagate the useful error message.

Result:

Show useful error messages are shown to the end user.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Fixes: https://github.com/dCache/dcache-view/issues/159
Acked-by: Lea Morschel

Reviewed at https://rb.dcache.org/r/11752/