dcache-view (frontend and webdav): encode file path which are part of url
Motivation:

Some operations in the restful api requires that the file path
be included in the url. If this url (that is,  the file path
part of the url) is not encoded, this will result in all sort
of errors such as the one reported at #164.

This often happens when the file path contains character/s that
are considered special.

Modification:

Create a method that take a file path and encode it. Use this
method where necessary.

Result:

Fixed issue #164 and all errors caused by unencoded url.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Acked-by: Lea Morschel
Acked-by: Paul Millar
Fixes: https://github.com/dCache/dcache-view/issues/164

Reviewed at https://rb.dcache.org/r/11785/