dcache-view: increase display time of the notification message
Motivation:

When the notification message is displayed it is shown for
a very short time and at times the user were not able to
read the message before it is close.

Modification:

- increase the display time for the toast
- create a function that close the current displayed toast
    before showing a new one.
- centralise how the toast is open by relying on
    dispatchEvent which make dcache-view behaviour consistent

Result:

display notification message for a longer period of time.

Target: master
Request: 1.5
Require-notes: no
Require-book: no
Acked-by: Paul Millar
Acked-by: Albert Rossi
Acked-by: Lea Morschel

Reviewed at https://rb.dcache.org/r/11792/