dcache-view: fix the drop-down qos transition choices
Motivation:

A small bug in the way the drop-down choices
for possible transitions are activated.

Modification:

Instead of short-circuiting the 'true' return
value so that it returns true if the
first item in the list does not match, one should
return false if a match is found, and true
only if it has iterated through all without
a match.  Since that is essentially
the 'contains' loop, we have used the built-in
javascript Array.includes.

Also added 'Volatile' to the QoS choices for
the list-row qosValue.

Result:

Correct logic.

Target: master
Request: 1.5
Requires-notes: yes
Requires-book: no
Acked-by: Paul
Acked-by: Lea