dcache-view (SSE): add inotify events subscription
Motivation:

After creating an event channel and watching it, the next
logical thing is to subscribe to events. At the moment,
the only interesting events the user can subscribe to is
the inotify event. This patch ensure that a inotify event
is subscribe to when a user is viewing a directory.

Modification:

- create a worker that can be use for subscription and
    cancelation of events.
- add utilities functions that makes use of the worker
    for subscribing and cancelling of events.
- subscribe to an event when a directory is currently
    in view.

Result:

Subscribe to inotify event based on the directory that
is currently being view.

Target: master
Requires-notes: no
Requires-book: yes
Acked-by: Tigran Mkrtchyan
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/12121/