dcache-view (shared-file): fix minor issue with drag and drop
Motivation:

Dnd in shared-files route is not working as expected
because the macaroon value is passed along and the
dnd listeners are not registered.

Modification:

- ensure that the macaroon auth value is propagated.
- add dnd event listners to shared-files route

Result:

Issue with dnd in the shared-files route is fixed and
it now work as expected.

Target: master
Request: 1.5
Requires-notes: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/12172/