dcache-view (dv.js): fix TypeError issue
Motivation:

When listing the content of a directory, the element
responsible for file listing is created dynamically
and appended to the right parent. The parent node
depends on whether the route is home or shared-file.

Previously, the method responsible for dir listing,
will just search for the view-file in the current view.
This is then removed and new one is append based on
the new path. Unfortunately, the search can return null
hence resulting to TypeError.

Modification:

If the view-file search return null, use a different
strategy for finding the parent node.

Result:

Avoid TypeError

Target: master
Request: 1.5
Requires-notes: no
Acked-by: Paul Millar

Reviewed at https://rb.dcache.org/r/12179/