dcache-view: adjust pool plots viewport
Motivation:

The admin pages share html segments with the other pages;
the particular host page fits into the bottom portion of
the total page.   In order to be able to display the
contents clearly without cutting off the edge, a
sweet-spot for the vertical dimension needs to be found.

Currently the pool plots use a dom-repeat distributed
as a grid inside a scrolling header area.  That area
is set to have 70% height.  The host viewport however
is set larger (80vh) in terms of the actual size of the
screen.  This results in cutting off the bottom of the
grid if there are not enough elements to scroll.

Modification:

This is a tricky bit of business, but some amelioration
is purchased by making the viewport vertically smaller
(70vh).  There may still be a need to adjust the browser
dimensions if its width is too long in relation to
height, but on a 13" Macbook screen as well as a
large panel screen, the link at the bottom now
shows without problem. (see attached).

Result:

Hopefully less complaints from users.

Target: master
Patch: https://rb.dcache.org/r/12707/
Requires-notes: yes (fixes a reported issue)
Acked-by:  Tigran